+++ mapping/point_cloud_mapping/src/cloud_io/tools/pcd_to_msg.cpp
@@ -90,7 +90,7 @@ class PCDGenerator
       {
         usleep (1000000);
 
-        ROS_INFO ("Publishing data (%d points) on topic %s.", msg_cloud_.pts.size (), node_.mapName (cloud_topic_).c_str ());
+        ROS_INFO ("Publishing data (%d points) on topic %s.", (int)msg_cloud_.pts.size (), node_.mapName (cloud_topic_).c_str ());
         msg_cloud_.header.stamp = ros::Time::now ();
         node_.publish ("cloud_pcd", msg_cloud_);
       }
@@ -98,7 +98,7 @@ class PCDGenerator
       return (true);
     }
 
-  
+
 };
 
 /* ---[ */
@@ -110,11 +110,11 @@ int
     ROS_ERROR ("Need one PCD file as parameter!");
     return (-1);
   }
-  
+
   ros::init (argc, argv);
 
   ros::Node ros_node ("pcd_generator");
-  
+
   PCDGenerator c (ros_node);
   c.file_name_ = string (argv[1]);
   ROS_INFO ("Loading file %s...", c.file_name_.c_str ());
